+++ src/main/java/io/hhplus/conbook/interfaces/schedule/token/TokenScheduler.java
@@ -1,26 +1,18 @@
-package io.hhplus.conbook.config;
+package io.hhplus.conbook.interfaces.schedule.token;
 
-import io.hhplus.conbook.domain.booking.BookingService;
 import io.hhplus.conbook.domain.concert.ConcertService;
 import io.hhplus.conbook.domain.token.TokenManager;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.TaskScheduler;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import java.time.Instant;
-import java.time.LocalDateTime;
-import java.time.ZoneId;
-
-@Component
 @Slf4j
+@Component
 @RequiredArgsConstructor
-public class ScheduledTaskExecutor {
+public class TokenScheduler {
     private final TokenManager tokenManager;
-    private final BookingService bookingService;
     private final ConcertService concertService;
-    private final TaskScheduler taskScheduler;
 
     /**
      * scheduler interval time : 5min
@@ -37,18 +29,5 @@ public class ScheduledTaskExecutor {
             tokenManager.clearNonValidTokens(concertId);
             tokenManager.convertToPass(concertId);
         }
-    }
-
-    public void addSchedule(long bookingId, long intervalMin) {
-        // 단발성 스케쥴러 등록
-        Runnable task = () -> {
-            log.info("\nTaskScheduler has been executed");
-
-            bookingService.checkOrUpdate(bookingId);
-        };
-        Instant startTime =
-                LocalDateTime.now().plusMinutes(intervalMin).atZone(ZoneId.systemDefault()).toInstant();
-
-        taskScheduler.schedule(task,startTime);
     }
 }
